# [Enhancement] Support execute-only flag for performance test