## Why I'm doing:

## What I'm doing:
don't send sql result to client, for test accurate performance

```
MySQL td> set enable_execution_only=false
Query OK, 0 rows affected
Time: 0.021s
MySQL td> select * from t1;
+--------+--------+-----------+
| v1     | v2     | a1        |
+--------+--------+-----------+
| <null> | <null> | ["a"]     |
| 1      | 1      | ["1","2"] |
| 1      | <null> | []        |
| 2      | 2      | ["3","4"] |
| 2      | <null> | []        |
| 3      | <null> | []        |
| 4      | <null> | []        |
| 5      | <null> | <null>    |
+--------+--------+-----------+
8 rows in set
Time: 0.099s
MySQL td> set enable_execution_only=true
Query OK, 0 rows affected
Time: 0.004s
MySQL td> select * from t1;
+----+----+----+
| v1 | v2 | a1 |
+----+----+----+
+----+----+----+
0 rows in set
Time: 0.059s
MySQL td>
```


Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5